[FIX]pms_fastapi: Move generic endpoints from roomdoo_fastapi to pms_fastapi
