+++ roomdoo_fastapi/routers/__init__.py
@@ -2,10 +2,6 @@ from . import user
 from . import instance
 from . import login
 from . import pms_property_url
-from . import customer
-from . import guest
-from . import agency
-from . import supplier
 from . import contact
 from . import zip_autocomplete
 from . import reports
