+++ roomdoo_fastapi/schemas/__init__.py
@@ -1,8 +1,5 @@
 from . import user
 from . import instance
-from . import customer
 from . import property_link
-from . import guest
-from . import supplier
 from . import contact
 from . import zip_autocomplete
