+++ roomdoo_fastapi/tests/__init__.py
@@ -1,8 +1,4 @@
-from . import test_agencies
-from . import test_customers
-from . import test_guests
 from . import test_instance
 from . import test_pms_property_urls
-from . import test_suppliers
 from . import test_user
 from . import test_contacts
